server/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2 
spacy
pthon-dotenv
pandas
openai
fastapi-cors